pyproject.toml
[project]
name = "company-scraper-agent"
version = "1.0.0"
description = "AI-powered company information scraping agent - 100% FREE"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "ddgs>=9.0.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "lxml>=5.0.0",
    "html5lib>=1.1",
    "langgraph>=0.2.0",
    "langchain>=0.2.0",
    "langchain-community>=0.2.0",
    "langchain-groq>=0.1.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "playwright>=1.40.0",
]

[project.optional-dependencies]
ollama = ["langchain-ollama>=0.1.0"]
dev = ["pytest>=8.0.0", "black>=24.0.0", "ruff>=0.3.0"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["agent*"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "W"]
